create like

+++ application/models/Photo.php
@@ -27,16 +27,14 @@ class Photo extends Model {
 
             if (empty($errors)) {
                 $fileNameNew = uniqid('', true). '.' . $fileExt;
-                $targetDir = 'public/images/' . $fileNameNew;
+                $targetDir = 'public/images/gallery/' . $fileNameNew;
 
                 $params = [
                     'image_id' =>null,
                     'user_id' => $_SESSION['account']['user_id'],
-                    'path' => '',
+                    'path' => '/' . $targetDir,
                 ];
                 $this->db->query('INSERT INTO gallery VALUES (:image_id, :user_id, NOW(), :path)', $params);
-                $lastId = $this->db->lastInsertId();
-                echo $lastId;
                 move_uploaded_file($fileTmpName, $targetDir);
                 echo 'success';
             } else {
@@ -49,30 +47,63 @@ class Photo extends Model {
 
     public function displayGallery() {
 
-//        $params = [
-//            'login' => $login,
-//        ];
+        return $this->db->row("SELECT * FROM `gallery` ORDER BY `user_id`");
+    }
 
-        $data = $this->db->row("SELECT * FROM `gallery` ORDER BY `user_id`");
-        echo "<form action='/photo/gallery' method='post'";
-        foreach ($data as $key => $val) {
-            $image = '<img src="' . $val['path'] . '" width="" height="150">';
-            $hidden = '<input type="hidden" name="path" value="' . $val['path'] . '">';
-            $delete = '<input type="submit" name="delete" value="delete">';
-            echo $image;
-            echo $hidden;
-            echo $delete;
+    public function checkLike($imageId) {
+        $params = [
+            'image_id' => $imageId,
+        ];
+        $data = $this->db->row('SELECT `user_id` FROM `likes` WHERE image_id = :image_id', $params);
+        foreach ($data as $val) {
+            if ($val['user_id'] == $_SESSION['account']['user_id']) {
+                return false;
+            }
         }
-        echo "</form>";
+        return true;
+    }
+
+    public function addLike($imageId) {
+        $params = [
+            'like_id' => null,
+            'user_id' => $_SESSION['account']['user_id'],
+            'image_id' => $imageId,
+        ];
+        $this->db->query('INSERT INTO `likes` VALUES (:like_id, :user_id, :image_id)', $params);
+    }
+
+    public function deleteLike($imageId) {
+        $params = [
+            'image_id' => $imageId,
+        ];
+        $data = $this->db->row('SELECT `user_id`, `like_id` FROM `likes` WHERE image_id = :image_id', $params);
+        foreach ($data as $val) {
+            if ($val['user_id'] == $_SESSION['account']['user_id']) {
+                $params = [
+                    'like_id' => $val['like_id'],
+                ];
+                $this->db->row("DELETE FROM `likes` WHERE like_id = :like_id", $params);
+            }
+        }
+    }
+
+    public function countLike($imageId) {
+        $params = [
+            'image_id' => $imageId,
+        ];
+        $data = $this->db->column('SELECT COUNT(*) FROM `likes` WHERE image_id = :image_id', $params);
+        echo $data;
     }
 
     public function deletePhoto($path) {
         $params = [
-            'id' => $path,
+            'path' => $path,
         ];
-        //echo $path;
-        $this->db->row("DELETE FROM `gallery` WHERE id = :id", $params);
-        unlink($path);
+        $this->db->row("DELETE FROM `gallery` WHERE path = :path", $params);
+        $path = trim($path, '/');
+        if (file_exists($path)) {
+            unlink($path);
+        }
     }
 
 }
